chore: 커뮤니티 카테고리 툴팁 코드 분리
